requirements.txt
python-telegram-bot==20.7
redis==5.0.1
python-dotenv==1.0.0
aiohttp==3.9.1
structlog==24.1.0
